ConnAnchor now uses float, but labels aren't inserted so tests are failing

+++ kite9-visualization/src/common/kotlin/org/kite9/diagram/visualization/compaction2/anchors/ConnAnchor.kt
@@ -6,5 +6,5 @@ import org.kite9.diagram.visualization.compaction2.anchors.Anchor
 /**
  * Anchor for a point in a connection, numbered from zero
  */
-data class ConnAnchor(override val e: Connection, override val s: Int) : Anchor<Int> {
+data class ConnAnchor(override val e: Connection, override val s: Float) : Anchor<Float> {
 }
